requirements.txt
torch
torchvision
pillow
tqdm
pyyaml
transformers
insightface
onnxruntime-gpu
dotenv
streamlit